made hazard domain in services
